+++ src/classes_body_ships.py
@@ -10,7 +10,6 @@ from classes_body_vehicles import *
 class Small_ship(Vehicle):
     path = SMALL_SHIP_PATH
     number_of_frames = SMALL_SHIP_FRAMES
-    number_of_frames_in_sequence = SMALL_SHIP_FRAMES - 1
     min_scale_to_be_visible = 0.5
 
     v_max = 0.5
@@ -29,7 +28,6 @@ class Small_ship(Vehicle):
 class Medium_ship(Small_ship):
     path = MEDIUM_SHIP_PATH
     number_of_frames = MEDIUM_SHIP_FRAMES
-    number_of_frames_in_sequence = MEDIUM_SHIP_FRAMES - 1
     min_scale_to_be_visible = 0.25
 
     v_max = 0.4
@@ -43,7 +41,6 @@ class Medium_ship(Small_ship):
 class Destroyer_body(Small_ship):
     path = DESTROYER_PATH
     number_of_frames = DESTROYER_FRAMES
-    number_of_frames_in_sequence = DESTROYER_FRAMES - 1
     min_scale_to_be_visible = 0.125
 
     v_max = 0.3
@@ -57,7 +54,6 @@ class Destroyer_body(Small_ship):
 class Battleship_body(Small_ship):
     path = BATTLESHIP_PATH
     number_of_frames = BATTLESHIP_FRAMES
-    number_of_frames_in_sequence = BATTLESHIP_FRAMES - 1
     min_scale_to_be_visible = 0.125
 
     v_max = 0.2